Moved RegisterResponse type
