+++ src/api/register.ts
@@ -1,18 +1,6 @@
 import axios, { AxiosError } from "axios";
+import { RegisterResponse } from "types/auth";
 import { baseUrl } from "./config";
-
-export type RegisterResponse = {
-  error: string;
-  validationErrors: null,
-} | {
-  validationErrors: {
-    [key: string]: string;
-  };
-  error: null
-} | {
-  error: null;
-  validationErrors: null;
-};
 
 export async function register(token: string, data: FormData): Promise<RegisterResponse> {
   try {
