+++ src/types/auth.ts
@@ -1,0 +1,12 @@
+export type RegisterResponse = {
+  error: string;
+  validationErrors: null,
+} | {
+  validationErrors: {
+    [key: string]: string;
+  };
+  error: null
+} | {
+  error: null;
+  validationErrors: null;
+};
